Move binary_operation to internal decorators
