+++ metafunctions/core/_base.py
@@ -1,8 +1,10 @@
-import functools
 import operator
-from collections.abc import Callable
 import typing as tp
 import abc
+import functools
+
+
+from metafunctions.core._decorators import binary_operation
 
 
 class MetaFunction(metaclass=abc.ABCMeta):
@@ -35,17 +37,6 @@ class MetaFunction(metaclass=abc.ABCMeta):
         '''Wrap the given value in a DeferredValue object, (which returns its value when called).'''
         return DeferredValue(value)
 
-    def _binary_operation(method):
-        '''Internal decorator to apply common type checking for binary operations'''
-        @functools.wraps(method)
-        def binary_operation(self, other):
-            if isinstance(other, Callable):
-                new_other = self.make_meta(other)
-            else:
-                new_other = self.defer_value(other)
-            return method(self, new_other)
-        return binary_operation
-
     def _modify_kwargs(self, kwargs: dict):
         '''Do pre-call modifications to the kwargs dictionary. Currently this just means adding a
         meta object if _bind is true.
@@ -53,48 +44,46 @@ class MetaFunction(metaclass=abc.ABCMeta):
         kwargs.setdefault('meta', self)
 
     ### Operator overloads ###
-    @_binary_operation
+    @binary_operation
     def __or__(self, other):
         return FunctionChain.combine(self, other)
 
-    @_binary_operation
+    @binary_operation
     def __ror__(self, other):
         return FunctionChain.combine(other, self)
 
-    @_binary_operation
+    @binary_operation
     def __add__(self, other):
         return FunctionMerge(operator.add, (self, other))
 
-    @_binary_operation
+    @binary_operation
     def __radd__(self, other):
         return FunctionMerge(operator.add, (other, self))
 
-    @_binary_operation
+    @binary_operation
     def __sub__(self, other):
         return FunctionMerge(operator.sub, (self, other))
 
-    @_binary_operation
+    @binary_operation
     def __rsub__(self, other):
         return FunctionMerge(operator.sub, (other, self))
 
-    @_binary_operation
+    @binary_operation
     def __mul__(self, other):
         return FunctionMerge(operator.mul, (self, other))
 
-    @_binary_operation
+    @binary_operation
     def __rmul__(self, other):
         return FunctionMerge(operator.mul, (other, self))
 
-    @_binary_operation
+    @binary_operation
     def __truediv__(self, other):
         return FunctionMerge(operator.truediv, (self, other))
 
-    @_binary_operation
+    @binary_operation
     def __rtruediv__(self, other):
         return FunctionMerge(operator.truediv, (other, self))
 
-    # This is almost definitely a bad idea, but it's interesting that it works
-    del _binary_operation
 
 
 class FunctionChain(MetaFunction):
